plot.power.htest {pwr}	R Documentation

Plot diagram of sample size vs. test power

Description

Plot a diagram to illustrate the relationship of sample size and test power for a given set of parameters.

Usage

## S3 method for class 'power.htest'
plot(x, ...)

Arguments

x
object of class power.htest usually created by one of the power calculation functions, e.g., pwr.t.test()

...
Arguments to be passed to ggplot including xlab and ylab

Details

Power calculations for the following tests are supported: t-test (pwr.t.test(), pwr.t2n.test()), chi squared test (pwr.chisq.test()), one-way ANOVA (pwr.anova.test(), standard normal distribution (pwr.norm.test()), Pearson correlation (pwr.r.test()), proportions (pwr.p.test(), pwr.2p.test(), pwr.2p2n.test()))

Value

These functions are invoked for their side effect of drawing on the active graphics device.

Note

By default it attempts to use the plotting tools of ggplot2 and scales. If they are not installed, it will use the basic R plotting tools.

Author(s)

Stephan Weibelzahl <weibelzahl@pfh.de>

See Also

pwr.t.test, pwr.p.test, pwr.2p.test, pwr.2p2n.test, pwr.r.test, pwr.chisq.test, pwr.anova.test, pwr.t2n.test

Examples

## Two-sample t-test
p.t.two <- pwr.t.test(d=0.3, power=0.8, type="two.sample", alternative="two.sided")
plot(p.t.two)
plot(p.t.two, xlab="sample size per group")

[Package pwr version 1.3-0 Index]