pwr.r.test {pwr}	R Documentation

Power calculations for correlation test

Description

Compute power of test or determine parameters to obtain target power (same as power.anova.test).

Usage

pwr.r.test(n = NULL, r = NULL, sig.level = 0.05, power = NULL,
    alternative = c("two.sided", "less","greater"))

Arguments

n
Number of observations

r
Linear correlation coefficient

sig.level
Significance level (Type I error probability)

power
Power of test (1 minus Type II error probability)

alternative
a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less"

Details

These calculations use the Z' transformation of correlation coefficient : Z'=arctanh(r) and a bias correction is applied. Note that contrary to Cohen (1988) p.546, where zp' = arctanh(rp) + rp/(2*(n-1)) and zc' = arctanh(rc) + rc/(2*(n-1)), we only use here zp' = arctanh(rp) + rp/(2*(n-1)) and zc' = arctanh(rc).

Exactly one of the parameters 'r','n','power' and 'sig.level' must be passed as NULL, and that parameter is determined from the others. Notice that the last one has non-NULL default so NULL must be explicitly passed if you want to compute it.

Value

Object of class '"power.htest"', a list of the arguments (including the computed one) augmented with 'method' and 'note' elements.

Note

'uniroot' is used to solve power equation for unknowns, so you may see errors from it, notably about inability to bracket the root when invalid arguments are given.

Author(s)

Stephane Champely <champely@univ-lyon1.fr> but this is a mere copy of Peter Dalgaard work (power.t.test). The modified bias correction is contributed by Jeffrey Gill.

References

Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.). Hillsdale,NJ: Lawrence Erlbaum.

Examples

## Exercise 3.1 p. 96 from Cohen (1988)
pwr.r.test(r=0.3,n=50,sig.level=0.05,alternative="two.sided")
pwr.r.test(r=0.3,n=50,sig.level=0.05,alternative="greater")

## Exercise 3.4 p. 208
pwr.r.test(r=0.3,power=0.80,sig.level=0.05,alternative="two.sided")
pwr.r.test(r=0.5,power=0.80,sig.level=0.05,alternative="two.sided")
pwr.r.test(r=0.1,power=0.80,sig.level=0.05,alternative="two.sided")

[Package pwr version 1.3-0 Index]